+++ craft_application/git/_git_repo.py
@@ -710,6 +710,21 @@ class GitRepo:
             ) from error
         return cls(path)
 
+    def get_config_value(self, key: str) -> str | None:
+        """Get value for the configuration key if available else return None."""
+        try:
+            return cast(str, self._repo.config[key])  # pyright: ignore[reportUnnecessaryCast]
+        except (KeyError, ValueError):
+            logger.debug("Config key %r not found in the repository", key)
+            return None
+
+    def set_config_value(self, key: str, value: str) -> None:
+        """Set new value for the configuration key.
+
+        :raises ValueError: if configuration is incorrect
+        """
+        self._repo.config[key] = value
+
     @classmethod
     @lru_cache(maxsize=1)
     def get_git_command(cls) -> str: