+++ tests/integration/git/test_git.py
@@ -16,10 +16,16 @@
 """Git module integration tests."""
 
 import pathlib
+import re
 
 import pytest
 
 from craft_application.git import NO_PUSH_URL, Commit, GitError, GitRepo
+
+
+@pytest.fixture
+def git_repo(empty_repository: pathlib.Path) -> GitRepo:
+    return GitRepo(empty_repository)
 
 
 @pytest.mark.slow
@@ -97,3 +103,49 @@ def test_set_no_push(
     git_repo.set_no_push(test_remote)
     assert git_repo.get_remote_url(remote_name=test_remote) == hello_repository_lp_url
     assert git_repo.get_remote_push_url(remote_name=test_remote) == NO_PUSH_URL
+
+
+@pytest.mark.parametrize(
+    ("config_key", "expected_value"),
+    [("core.bare", "false"), ("non_existent.key", None)],
+)
+def test_get_repo_configuration(
+    git_repo: GitRepo, config_key: str, expected_value: str | None
+) -> None:
+    assert git_repo.get_config_value(config_key) == expected_value
+
+
+def test_set_repo_configuration(git_repo: GitRepo) -> None:
+    new_key = "test.craft"
+    new_value = "just-testing"
+    assert git_repo.get_config_value(new_key) is None
+    git_repo.set_config_value(new_key, new_value)
+    assert git_repo.get_config_value(new_key) == new_value
+
+
+def test_update_repo_configuration(git_repo: GitRepo) -> None:
+    key = "test.craft"
+    old_value = "just-old"
+    new_value = "just-new"
+    git_repo.set_config_value(key, old_value)
+    assert git_repo.get_config_value(key) == old_value
+    git_repo.set_config_value(key, new_value)
+    assert git_repo.get_config_value(key) == new_value
+
+
+def test_update_boolean_with_string_value(git_repo: GitRepo) -> None:
+    key = "core.bare"
+    new_value = "incorrect-boolean"
+    # this makes repository inaccessible via git CLI client
+    # fatal: bad boolean config value 'incorrect-boolean' for 'core.bare'
+    git_repo.set_config_value(key, new_value)
+    assert git_repo.get_config_value(key) == new_value
+
+
+def test_incorrect_config_key(git_repo: GitRepo) -> None:
+    key = "craft.incorrect&test*key"
+    new_value = "not-important"
+    with pytest.raises(
+        ValueError, match=re.escape(f"invalid config item name {key!r}")
+    ):
+        git_repo.set_config_value(key, new_value)