+++ docs/reference/changelog.rst
@@ -18,13 +18,18 @@ Commands
 
 - Reduce spread verbosity level when running the ``test`` command.
 
+Git
+===
+
+- Add API to modify repository configuration.
+
 Fixes
 ======
 
 - Improve test result messages.
 - ``InitService`` no longer leaves empty files if rendering template fails.
 
-For a complete list of commits, check out the `5.2.1`_ release on GitHub.
+For a complete list of commits, check out the `5.3.0`_ release on GitHub.
 
 5.2.0 (2025-04-25)
 ------------------
@@ -772,4 +777,4 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _5.0.4: https://github.com/canonical/craft-application/releases/tag/5.0.4
 .. _5.1.0: https://github.com/canonical/craft-application/releases/tag/5.1.0
 .. _5.2.0: https://github.com/canonical/craft-application/releases/tag/5.2.0
-.. _5.2.1: https://github.com/canonical/craft-application/releases/tag/5.2.1
+.. _5.3.0: https://github.com/canonical/craft-application/releases/tag/5.3.0